Moving change-password to myprofile level
Change password feature was on the profile control panel but was related
to the user logged in. This was an inconsistent behavior since all the
actions on the control panel are related to the profile. So I moved the
change password feature to a new controller on myprofile and now fixed
so that you change the profile password instead of the user logged in.

(ActionItem2025)